Created  routers using `createBrowserRouter` from the `react-router-dom` library.
It is creating a
router using `createBrowserRouter` from the `react-router-dom` library. The router is configured
with different paths and corresponding components to render. The `ReactDOM.createRoot` function is
used to render the application, providing the router as a prop to the `RouterProvider` component.
The application is rendered inside the element with the id "root" in the HTML document.